requirements.txt
selenium==4.16.0
undetected-chromedriver==3.5.4
pandas==2.1.3
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
